sar_fetcher/requirements.txt
sentinelhub>=3.9
tqdm>=4.65.0
pandas>=2.0.0
geopandas>=0.13.0
click>=8.1.0
python-dateutil>=2.8.2
